move sink
